analitica_de_compras/requirements.txt
confluent_kafka==2.10.1
attrs==25.3.0
httpx==0.28.1
jsonschema==4.24.0
referencing==0.36.2
Authlib==1.6.0
cachetools==6.1.0
python-dotenv==0.20.0
avro==1.12.0
urllib3==2.5.0
requests==2.32.5
pyspark==4.1.0
Django==3.2.16
djangorestframework==3.12.4
